+++ src/main/java/com/farcr/nomansland/common/blockentity/MonsterAnchorBlockEntity.java
@@ -3,9 +3,9 @@ package com.farcr.nomansland.common.blockentity;
 import com.farcr.nomansland.NMLConfig;
 import com.farcr.nomansland.common.block.MonsterAnchorBlock;
 import com.farcr.nomansland.common.event.listener.AnchorListener;
+import com.farcr.nomansland.common.registry.NMLBlockEntities;
 import com.farcr.nomansland.common.registry.NMLParticleTypes;
 import com.farcr.nomansland.common.registry.NMLSounds;
-import com.farcr.nomansland.common.registry.blocks.NMLBlockEntities;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.particles.ParticleOptions;
