+++ src/main/java/com/farcr/nomansland/common/blockentity/TapBlockEntity.java
@@ -2,7 +2,7 @@ package com.farcr.nomansland.common.blockentity;
 
 import com.farcr.nomansland.NMLConfig;
 import com.farcr.nomansland.common.block.cauldrons.FourLayeredCauldronBlock;
-import com.farcr.nomansland.common.registry.blocks.NMLBlockEntities;
+import com.farcr.nomansland.common.registry.NMLBlockEntities;
 import com.farcr.nomansland.common.registry.blocks.NMLBlocks;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.particles.ParticleTypes;
