+++ src/main/java/com/farcr/nomansland/common/integration/FDIntegration.java
@@ -1,5 +1,6 @@
 package com.farcr.nomansland.common.integration;
 
+import com.farcr.nomansland.common.registry.blocks.BlockDefinition;
 import com.farcr.nomansland.common.registry.blocks.NMLBlocks;
 import com.farcr.nomansland.common.registry.items.NMLFoods;
 import com.farcr.nomansland.common.registry.items.NMLItems;
@@ -9,7 +10,6 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.neoforged.neoforge.event.BlockEntityTypeAddBlocksEvent;
-import net.neoforged.neoforge.registries.DeferredBlock;
 import net.neoforged.neoforge.registries.DeferredItem;
 import vectorwing.farmersdelight.common.FoodValues;
 import vectorwing.farmersdelight.common.block.CabinetBlock;
@@ -20,25 +20,26 @@ import vectorwing.farmersdelight.common.item.DrinkableItem;
 import vectorwing.farmersdelight.common.item.MushroomColonyItem;
 import vectorwing.farmersdelight.common.registry.ModBlockEntityTypes;
 
-import static com.farcr.nomansland.common.registry.blocks.NMLBlocks.*;
+import static com.farcr.nomansland.common.registry.blocks.NMLBlocks.registerBlock;
+import static com.farcr.nomansland.common.registry.blocks.NMLBlocks.registerBlockNoItem;
 import static com.farcr.nomansland.common.registry.items.NMLItems.registerItem;
 import static net.minecraft.world.level.block.state.BlockBehaviour.Properties.ofFullCopy;
 
 public class FDIntegration {
 
-    public static final NMLBlocks.BlockDefinition<Block> PINE_CABINET = registerBlock("pine_cabinet",
+    public static final BlockDefinition<Block> PINE_CABINET = registerBlock("pine_cabinet",
             () -> new CabinetBlock(ofFullCopy(Blocks.BARREL)));
 
-    public static final NMLBlocks.BlockDefinition<Block> MAPLE_CABINET = registerBlock("maple_cabinet",
+    public static final BlockDefinition<Block> MAPLE_CABINET = registerBlock("maple_cabinet",
             () -> new CabinetBlock(ofFullCopy(Blocks.BARREL)));
 
-    public static final NMLBlocks.BlockDefinition<Block> WALNUT_CABINET = registerBlock("walnut_cabinet",
+    public static final BlockDefinition<Block> WALNUT_CABINET = registerBlock("walnut_cabinet",
             () -> new CabinetBlock(ofFullCopy(Blocks.BARREL)));
 
-    public static final NMLBlocks.BlockDefinition<Block> WILLOW_CABINET = registerBlock("willow_cabinet",
+    public static final BlockDefinition<Block> WILLOW_CABINET = registerBlock("willow_cabinet",
             () -> new CabinetBlock(ofFullCopy(Blocks.BARREL)));
 
-    public static final NMLBlocks.BlockDefinition<Block> FIELD_MUSHROOM_COLONY = registerBlockNoItem("field_mushroom_colony",
+    public static final BlockDefinition<Block> FIELD_MUSHROOM_COLONY = registerBlockNoItem("field_mushroom_colony",
             () -> new MushroomColonyBlock(NMLItems.FIELD_MUSHROOM, Block.Properties.ofFullCopy(NMLBlocks.FIELD_MUSHROOM.get())));
 
     public static final DeferredItem<Item> FIELD_MUSHROOM_COLONY_ITEM = registerItem("field_mushroom_colony",
@@ -50,7 +51,7 @@ public class FDIntegration {
     public static final DeferredItem<Item> PEAR_COBBLER_SLICE = registerItem("pear_cobbler_slice",
             () -> new Item(new Item.Properties().food(FoodValues.PIE_SLICE)));
 
-    public static final NMLBlocks.BlockDefinition<Block> PEAR_COBBLER = registerBlockNoItem("pear_cobbler",
+    public static final BlockDefinition<Block> PEAR_COBBLER = registerBlockNoItem("pear_cobbler",
             () -> new PieBlock(ofFullCopy(Blocks.CAKE), PEAR_COBBLER_SLICE));
 
     public static final DeferredItem<Item> PEAR_COBBLER_ITEM = registerItem("pear_cobbler",
