+++ src/main/java/com/farcr/nomansland/common/registry/worldgen/NMLFeatures.java
@@ -8,7 +8,6 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.configurations.NoneFeatureConfiguration;
-import net.minecraft.world.level.levelgen.feature.configurations.RandomPatchConfiguration;
 import net.neoforged.neoforge.registries.DeferredRegister;
 
 import java.util.function.Supplier;
