+++ src/main/java/com/farcr/nomansland/datagen/loot/FlowerPotBlockLootType.java
@@ -12,7 +12,7 @@ public class FlowerPotBlockLootType extends BlockLootType {
         this.plant = plant;
     }
 
-    ItemLike getPlant() {
+    public ItemLike getPlant() {
         return plant.get();
     }
 }
